pi/requirements.txt
absl-py==2.2.2
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-neopixel==6.3.15
adafruit-circuitpython-pixelbuf==2.0.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
astunparse==1.6.3
binho-host-adapter==0.1.6
certifi==2025.1.31
charset-normalizer==3.4.1
flatbuffers==20181003210633
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pigpio==1.78
protobuf==5.29.4
pyftdi==0.56.0
Pygments==2.19.1
pyserial==3.5
pyusb==1.3.1
requests==2.32.3
rich==14.0.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.17.0
sysv-ipc==1.1.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
